+++ bridgebot/game/bridgehand.py
