+++ bridgebot/game/card.py
