+++ bridgebot/game/enums.py
