+++ bridgebot/game/scoring.py
